python-engine/requirements.txt
torch==2.0.0
numpy==1.24.0
openai==1.3.0
python-dotenv==1.0.0
flask==2.3.0
flask-cors==4.0.0
opencv-python==4.8.0.74
pillow==10.0.0
rembg==2.0.50
colorsys
requests==2.31.0
